Add `showForSingleQuality` config to `qualitySelector`
Sometimes it makes sense to show the quality selector even if there is
only one quality to select from. This shows the user the current quality
for example. But it is also about consistency, so that the user can
always expect there to be this selector, otherwise users might be
confused by it missing. This is especially relevant if you move the
quality selector into a submenu: if the quality selector randomly
disappears, the submenu might end up empty, which looks very weird. So
in these cases, it is beneficial to just always show it.